feat(traces): keep closed trades' order traces in a local archive and resolve them through one path (#622)
The core answers `request_traces(ReportUID)` from an archive it keeps for a
bounded time, and MoonProto keeps no cache: the protocol asks the terminal to
store what it received by `ReportUID` and to ask only for what it does not
have. Until now the answers lived in `CoreData::report_traces` for the
session, the trade window asked the core on every open, and a trade the core
had aged out was gone for good.

- `data/order_traces.sqlite` (`db::order_traces`): one writer thread, tables
  `trace_answers` (per row: check time, line count) and `trace_lines` (points
  as f64 ms / f64 price, the archive is not downcast); an empty answer is
  filed with its date and re-asked only after 7 days; a failed request is
  never filed. The file joins the startup core-uid floor.
- the feed files every `TraceReady` in the archive, queues an ask when a
  report row closes (`RowUpsert` with `CloseDate != 0`, field indices cached
  per schema revision, the writer decides whether the row is already known)
  and, after a core's `SyncComplete`, backfills the closed rows of the last
  30 days without a trace, newest first, up to 400 per start
- every `request_traces` of a connection goes through one paced queue
  (`live::trace_backfill`): at most 8 unanswered, one send per 200 ms, a
  window's ask ahead of the backfill; three failures in a row abandon the
  queue (an old core never answers)
- one trace resolver on `Backend` (`backend::traces`, wake channel
  `TracesRevision`): archive read off-thread first, the core for misses up to
  a caller's cap, answers adopted from the session store on the feed drain
  edge (`DrainStats::report_traces`, which no longer repaints the backend's
  observers) and on the slow tick for a removed core
- the trade window consumes the resolver instead of polling the store itself;
  `ArchivedOrderTrace` prices become f64, narrowed to f32 only when a
  `LineTrace` is built

Verified on 26 live cores: 4620 backfill asks answered in 82 s at 5/s per
core with zero failures; a restart listed 0 rows for the 21 cores that were
fully filed and continued past the cap on the other 5.

Refs #621

+++ crates/moon-core/src/feed/report_traces.rs
@@ -30,12 +30,16 @@ pub struct ArchivedOrderTrace {
     pub kind: ArchivedLineKind,
     /// Stop line the core delivered with the trace (Moonbot `SetStopPrice`), horizontal at one
     /// price from the first point's time to `stop_time_ms`. `None` when the archive holds none.
-    pub stop_price: Option<f32>,
+    pub stop_price: Option<f64>,
     pub stop_time_ms: Option<f64>,
     /// `(Unix UTC ms, price)`, in the core's chart-geometry layout. Unset coordinates (a zero
     /// time or a non-positive price) are already dropped, and a trace that lost every point is
     /// not constructed at all.
-    pub points: Vec<(f64, f32)>,
+    ///
+    /// Prices stay the wire's `f64`: this type is what the local archive persists, and the
+    /// protocol asks that the saved archive not be downcast (`docs/reports.md`). The chart
+    /// narrows to `f32` at the moment it builds a `LineTrace`, not before.
+    pub points: Vec<(f64, f64)>,
 }
 
 /// What one `request_traces` asked for one report row ended in.
@@ -73,13 +77,13 @@ pub fn archived_traces_from_proto(traces: &[moonproto::ReportTrace]) -> Vec<Arch
                 | moonproto::OrderType::BuyLimit => ArchivedLineKind::Entry,
                 _ => return None,
             };
-            let points: Vec<(f64, f32)> = trace
+            let points: Vec<(f64, f64)> = trace
                 .points
                 .iter()
                 .filter_map(|p| {
                     let time_ms = p.time.unix_millis();
                     (time_ms > 1 && p.price.is_finite() && p.price > 0.0)
-                        .then_some((time_ms as f64, p.price as f32))
+                        .then_some((time_ms as f64, p.price))
                 })
                 .collect();
             if points.is_empty() {
@@ -91,7 +95,7 @@ pub fn archived_traces_from_proto(traces: &[moonproto::ReportTrace]) -> Vec<Arch
             Some(ArchivedOrderTrace {
                 own: trace.own,
                 kind,
-                stop_price: has_stop.then_some(trace.stop_price as f32),
+                stop_price: has_stop.then_some(trace.stop_price),
                 stop_time_ms: has_stop.then_some(stop_time_ms as f64),
                 points,
             })
